All functionality for User model is working, incuding BCrypt verification and authentication. Moved Home file into users views folder.

+++ app/controllers/beers_controller.rb
@@ -1,7 +1,12 @@
 class BeersController < ApplicationController
       
   def index
-    @beer = Beer.order(:name)
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @beer = Beer.order(:name)
+    else
+      redirect_to "/"
+    end
   end
   
   def new
@@ -9,31 +14,56 @@ class BeersController < ApplicationController
   end
   
   def create
-    @beer = Beer.new(params[:beer])
-    if @beer.save
-      redirect_to "/beers"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @beer = Beer.new(params[:beer])
+      if @beer.save
+        redirect_to "/beers"
+      else
+        render "new"
+      end
     else
-      render "new"
-    end
+      redirect_to "/"
+    end  
   end
   
   def show
-    @beer = Beer.find_by_id(params[:id])
-    @all_beer = Beer.where("name = '#{@beer.name}'")
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @beer = Beer.find_by_id(params[:id])
+      @all_beer = Beer.where("name = '#{@beer.name}'")
+    else
+      redirect_to "/"
+    end
   end
   
   def edit
-    @beer = Beer.find_by_id(params[:id])
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @beer = Beer.find_by_id(params[:id])
+    else
+      redirect_to "/"
+    end
   end
   
   def update
-    Beer.update(params[:id], params[:beer])
-    redirect_to "/beers"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      Beer.update(params[:id], params[:beer])
+      redirect_to "/beers"
+    else
+      redirect_to "/"
+    end
   end
   
   def destroy
-    Beer.delete_all("id = '#{params[:id]}'")
-    redirect_to "/beers"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      Beer.delete_all("id = '#{params[:id]}'")
+      redirect_to "/beers"
+    else
+      redirect_to "/"
+    end
   end
   
 end
