+++ app/controllers/breweries_controller.rb
@@ -11,32 +11,57 @@ class BreweriesController < ApplicationController
   end
   
   def create
-    @brewery = Brewery.new(params[:brewery])
-    if @brewery.save
-      redirect_to "/breweries"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @brewery = Brewery.new(params[:brewery])
+      if @brewery.save
+        redirect_to "/breweries"
+      else
+        render "new"
+      end
     else
-      render "new"
+      redirect_to "/"
     end
   end
   
   def show
-    beer_mapper
-    @brewery = Brewery.find_by_id(params[:id])
-    @beer_brewery = @brewery.beers
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      beer_mapper
+      @brewery = Brewery.find_by_id(params[:id])
+      @beer_brewery = @brewery.beers
+    else
+      redirect_to "/"
+    end
   end
   
   def edit
-    @brewery = Brewery.find_by_id(params[:id])
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      @brewery = Brewery.find_by_id(params[:id])
+    else
+      redirect_to "/"
+    end
   end
   
   def update
-    Brewery.update(params[:id], params[:brewery])
-    redirect_to "/breweries"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      Brewery.update(params[:id], params[:brewery])
+      redirect_to "/breweries"
+    else
+      redirect_to "/"
+    end
   end
   
   def destroy
-    Brewery.delete_all("id = '#{params[:id]}'")
-    redirect_to "/breweries"
+    if session[:user_id]
+      @current_user = User.find(session[:user_id])
+      Brewery.delete_all("id = '#{params[:id]}'")
+      redirect_to "/breweries"
+    else
+      redirect_to "/"
+    end
   end
   
 end
